Make the ALB health check unhealthy threshold configurable
Allow clients to set the number of health checks which need to fail before
an application is considered unhealthy.

This will let us raise the threshold for Jenkins, which is sometimes down
for a few minutes while restarting. In that case, it's better to wait to
see whether it starts up again before deploying a new ECS container.