Allow multiple CORS origins to upload to an S3 bucket
Update the parameters for the S3 module so that multiple frontend origins can be
passed in a parameter.

This will allow us to configure integration so that files can be uploaded from the
local dev environment as well as from the integration frontend, which makes it
possible to do local dev against a real integration S3 bucket.

If no origins are passed, do not enable CORS at all.