Remove async event config and DLQ from lambda.
The dead letter queue and retry policy within lambda doesn't work when
the event source in an sqs queue.

The relevant bit in the documentation is here:
https://docs.aws.amazon.com/lambda/latest/dg/invocation-retries.html

Event source mappings – Event source mappings that read from streams retry
the entire batch of items. Repeated errors block processing of the
affected shard until the error is resolved or the items expire.
To detect stalled shards, you can monitor the Iterator Age metric.

For event source mappings that read from a queue, you determine the length
of time between retries and destination for failed events by configuring
the visibility timeout and redrive policy on the source queue.