Increased the timeouts for the lambdas.
The checksum lambda processed a 3Gb file in just over two minutes so 3
seems like a reasonable compromise at this point.

I've increased the lambda to the same although it doesn't usually
appear to take this long. It keeps things consistent anyway and most
errors we get are memory or the code throwing exceptions so the number
of hanging lambdas should be negligable.