Add a lifecycle policy to remove untagged images.
When we push an image to ecr with a tag which is already assigned to an
image, intg or latest or whichever, the original image is kept in an
untagged state.
There's no reason for us to keep these. They don't affect our rollback
process at all so they're just taking up space.
This rule will delete them once the image is over a day old.