Pass EC2 instance ID instead of name to maintenance window module
This forces the module to wait for the EC2 instance to be created,
which means the maintenance task is run on the latest version of
the instance.

This fixes a bug where the EC2 instance was updated, but Terraform
couldn't tell that the instance ID needed to be updated to match.