Change modules to use output variables.
We were running into trouble trying to deploy this on staging because
some of the data blocks which linked to resources in another module
hadn't been created when terraform tried to use them.

This is because terraform can't work out the dependencies if you use
data sources in a separate module. It can do it if you output the data
you need from a module and pass it as a variable into another module so
this is what I've done.