Add image_source_url variable to ECR module
This parameter should be the URL of a Dockerfile or some other source
of the Docker images stored in the ECR repository.

This parameter would be helpful when someone needs to fix the Docker
image, e.g to fix a vulnerability found by an ECR image scan, but is
unfamiliar with the ECR repo. The URL will help them find the right
Dockerfile to build.

The default value of "unknown" will be removed once all the usages of
this module pass the parameter.