Allow management lambdas to use account kms key
These lambdas int the policy are created in the tdr-terraform-backend project and the
tdr-accounts project and so they can't use keys created in the
tdr-terraform-environments project as they won't exist yet if we're
running it from scratch.
So I'm using a key which the tdr-accounts project creates as this will
be run before the backend terraform.