Increase authoriser memory
The export api authoriser lambda is taking longer than 10 seconds since
we added the encrypted environment variables.
It's possible that this is caused by a high cpu load when it's
decrypting and as memory in lambda is linked to cpu, increasing the
memory should increase the cpu and speed it up a bit.