Only allow requests with correct host.
I've set this up so that the default behaviour of the load balancer
listener is to send a 403 to any request.
I then add a rule which sends requests through to the target group if
the header is set to the auth url.
This is working and it's stopped keycloak sending me a password reset
link for my-very-evil-service.com.